src/borzoi_GAME/src/predictor_script_and_utils/borzoi_API_script_and_utils/baskerville/pyproject.toml
[build-system]
requires = ["setuptools>=69.0.3", "setuptools_scm>=8.0.4"]
build-backend = "setuptools.build_meta"

[project]
name = "baskerville"
description = "baskerville"
authors = [
    {name = "Calico f(DNA)", email = "drk@calicolabs.com"},
]
readme = "README.md"
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version"]

requires-python = ">=3.9"
dependencies = [
    "h5py~=3.10.0",
    "intervaltree~=3.1.0",
    "joblib~=1.1.1",
    "matplotlib~=3.7.1",
    "google-cloud-storage~=2.0.0",
    "natsort~=7.1.1",
    "networkx~=2.8.4",
    "numpy~=1.24.3",
    "pandas~=1.5.3",
    "pybigwig~=0.3.18",
    "pybedtools~=0.10.0",
    "pysam~=0.22.0",
    "qnorm~=0.8.1",
    "seaborn~=0.12.2",
    "scikit-learn~=1.2.2",
    "scipy~=1.9.1",
    "statsmodels~=0.13.5",
    "tabulate~=0.8.10",
    "tensorflow~=2.15.0",
    "tqdm~=4.65.0",
    "pyfaidx~=0.7.1",
]

[project.optional-dependencies]
dev = [
    "black~=23.12.1",
    "pytest~=7.4.4",
    "ruff~=0.1.11",
]

gpu = [
    "tensorrt==8.6.1"
]

[project.urls]
Homepage = "https://github.com/calico/baskerville"

[tool.setuptools_scm]